test_queue_trigger/requirements.txt
azure-functions
azure-storage-queue == 12.14.0b1